+++ mayan/apps/file_metadata/drivers/__init__.py
@@ -1,1 +1,2 @@
 from .exiftool import *  # NOQA
+from .extract_msg import *  # NOQA